requirements.txt
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
flask
